Updated CI/CD workflows and project references
Removed the `dotnet-desktop.yml` and `AzurePipeline.yaml` files, which were GitHub Actions and Azure Pipelines workflows respectively for the Simple Encryptor Decryptor application. Added new `GitHubActions.yml` and `AzurePipeline.yml` files that perform similar tasks. These new workflows are triggered on push to the main branch and include jobs for parsing and validating the build version, building and publishing the project, and creating a GitHub release and tag. Updated the `SimpleEncryptorDecryptor.sln` file to remove the reference to the deleted `AzurePipeline.yaml` file and add references to the new `GitHubActions.yml` and `AzurePipeline.yml` files. The Visual Studio version in `SimpleEncryptorDecryptor.sln` remains unchanged at version 17.8.34511.84.