Update GitHub Actions badge in README
Updated the GitHub Actions workflow badge in the README.md file. The previous badge for the "Simple Encryptor Decryptor Production" workflow has been replaced with a new badge for the "Build & Release" workflow. This change reflects the current status of the workflow used for building and releasing the "Encryptor Decryptor v1.0" application.